[Fix][c] temperature should be correct now

+++ Arduino/sun_screen_module/getTemp.h
@@ -1,5 +1,5 @@
 
-int readTemp(void);
+float readTemp(void);
 void conversion(int);
 void analogToDigital(void);
 void initSensor(void);
